[UI 기능 추가] [Prefab 생성 시 오버라이트 여부 묻는 팝업 추가]-SceneOutliner 그 외 경고 문구들 일부 제거(형변환 및 문자열 관련) RenderManager 보기 좋게 함수화

+++ Project/Script/JugHand_Attack.cpp
@@ -243,7 +243,7 @@ void JugHand_Attack::Enter()
 
 	if (3 == iIndex)
 	{
-		float fHand3RunTime = m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime();
+		float fHand3RunTime = (float)(m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime());
 		fHand3RunTime *= 3.f;
 
 		GetOwner()->Animator3D()->Play(sAnimName, true);
@@ -254,7 +254,7 @@ void JugHand_Attack::Enter()
 	{
 		m_pAnimation->SetLerpTime(0.f);
 
-		float fRunTime = m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime();
+		float fRunTime = (float)(m_pAnimation->GetEndTime() - m_pAnimation->GetStartTime());
 
 		if (true == m_bFirstAttackDone)
 		{
